doc: move installation files to a common folder
This commit is part of the process of moving
the installation instructions from the website to
the docs.
As we need to exclude the open source-only files
in the Enterprise repository, the files that should not
be included must be moved to a common directory.
The name of the new directory is "installation-common".

In addition:
- The Configuration Reference file was moved out of the installation
  section (it's not about installation at all)
- The links to creating a cluster were removed from the installation
page (as not related).