sstables/storage: Add location() call
The call is returning a place where sstable or its component is.

For filesystem backed sstable "sstable location" is path to its Data
file, since historically it "represents" sstable in logs/tools/API.

For S3-backed sstable it returns a URL to the sstable files' prefix, or
a URL of the specific component's object. In rare case when sstable is
not yet "opened" and thus its UUID is not yet allocated it prints some
unaccessible URL representing sstable file-like name.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>