repair: Skip expired rows that do not shadow local data on the receiver
Consider the following example:

- Node n1, n2 in the cluster
- Create a table with RF 2
- Insert rows with 'using ttl 10'
- After the rows are expired
- Compaction runs on n1 or n2
- Run repair on n1

After the rows are expired, it is possible that compaction on n1 or n2
could purge the expired rows at different times. As a result, when
repair runs, it is possible that n1 has the expired data but n2 does
not. Repair will sync the expired data to n2 from n1. Node n1 could not
just ignore the expired rows, because the expired rows might shadow rows
on n2.

This patch improves the handling of the expired rows by checking on the
receiver node. If the receiver does not have any rows the expired row
might shadow, it skips writing the expired row to disk. This prevents
compaction and repair purging and bringing back the expired rows back
and forth. For example:

Before

- n1: expired rows, n2: expired rows
- compaction
- n1: expired rows, n2: no expired rows
- repair
- n1: expired rows, n2: expired rows
- compaction
- n1: no expired rows, n2: expired rows
- repair
- n1: expired rows, n2: expired rows

After

- n1: expired rows, n2: expired rows
- compaction
- n1: expired rows, n2: no expired rows
- repair
- n1: expired rows, n2: no expired rows
- compaction
- n1: no expired rows, n2: no expired rows
- repair
- n1: no expired rows, n2: no expired rows

Fixes #13308
Tests dtest/test_skip_expired_rows_on_repair