test.py: group non-boost tests by its suite name
before this change, the non-boost tests are grouped under the testsuite
named "non-boost tests", and their tests are flattened under this
test suite, so all shortname x mode x id combinations are listed at
the same level. this is not quite visually useful.
```xml
<testsuite name="non-boost tests" errors="0" tests="2" failures="0">
<testcase name="cql.locator_test.debug.1"/>
<testcase name="cql.lua_test.debug.1"/>
</testsuite>
````

after this change, the testcases are grouped under their individual
testsuites and the mode ("debug" in this case) and test.id ("1" in
this case) are encoded as the test's name. so the generated XML looks
like:
```xml
<testsuites>
<testsuite name="topology" tests="2" failures="0">
<testcase classname="test_mutation_schema_change" name="debug.1"/>
<testcase classname="test_mutation_schema_change" name="debug.2"/>
</testsuite>
</testsuites>
```

we could go further by matching the consolidated junit.xml with
each pytest's individual XML logger file, like how we consolidate
the boost::test's logger file, but let's leave it for another change.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>