compress: Add cache to named/id lookup of compressors
Minimize memory footprint by ensuring we re-use compressors even when
they are not in the default set.
As all compressors except zstd are 100% stateless, and the latter is only
stateful in that it has a few parameters, two of which are somewhat
useless, we cache only on name here, but add a "replace" mechanism
to allow some (zstd) to correctly set per-compressor compression level,
while still sharing the lions share of data structures.