test: make run_with_temporary_dir_pids a dict
before this change, run_with_temporary_dir_pids is a set of pids,
and we assume that the run_dir for each of the process in this
set are created by make_new_tempdir(), so all of the directories
are removed in cleanup_all(). but that's not always the case,
in the test of object_store, the run_dir is vendored by the test
itself, and the naming of the run_dir is not necessarily aligned
with the pid_to_dir. actually, the run_dir is created using
pid_to_pid(), but the "pid" here is the parent pid of the child
process created by run_with_generated_dir().

hence the assumption does not always necessarily hold. as we cannot
assume the behavior of run_dir_generator parameter passed to
run_with_generated_dir().

so, in this change,

* in addition to keeping the pid, we also keep the run_dir_generator
  in the run_with_temporary_dir_pids
* change the function signature of run_dir_generator(), so it
  also accepts a boolean indicating if we actually want to create
  the directory. this way, we can
  1. check if the directory is create by run.py or is vendored by
     the caller by checking if the generator function is the one
     created by make_new_tempdir() before removing it.
  2. open the log file in the correct path, instead of assuming
     that the run_dir is created usign pid_to_dir(pid).
* remove abort_run_with_temporary_dir(). this helper function is
  too small, so we just call the wrapped function directly.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>