storage_service: remove optimization in cleanup_group0_config_if_needed
The `topology_coordinator::cleanup_group0_config_if_needed` function
first checks whether the number of group 0 members is larger than the
number of non-left entries in the topology table, then attempts to
remove nodes in left state from group 0 and prints a warning if no such
nodes are found. There are some problems with this check:

- Currently, a node is added to group 0 before it inserts its entry to
  the topology table. Such a node may cause the check to succeed but no
  nodes will be removed, which will cause the warning to be printed
  needlessly.
- Cluster features on raft will reverse the situation and it will be
  possible for an entry in system.topology to exist without the
  corresponding node being a part of group 0. This, in turn, may cause
  the check not to pass when it should and nodes could be removed later
  than necessary.

This commit gets rid of the optimization and the warning, and the
topology coordinator will always compute the set of nodes that should be
removed. Additionally, the set of nodes to remove is now computed
differently: instead of iterating over left nodes and including only
those that are in group 0, we now iterate over group 0 members and
include those that are in `left` state. As the number of left nodes can
potentially grow unbounded and the number of group 0 members is more
likely to be bounded, this should give better performance in
long-running clusters.